Move context directory to shared/context
