+++ webapp/src/shared/hooks/UseAuth.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { AuthContext } from '../../context/auth-context';
+import { AuthContext } from '../context/auth-context';
 
 export function useAuth() {
   return useContext(AuthContext);
